refactor: useEscapeKeyClose 기능 분리
