+++ components/src/lib/Modal.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { useEffect } from 'react';
+import useEscapeKeyClose from './hooks/useEscapePress';
 
 interface ModalProps {
   isOpen: boolean;
@@ -7,19 +7,7 @@ interface ModalProps {
 }
 
 const Modal = ({ isOpen, onClose }: ModalProps) => {
-  useEffect(() => {
-    if (!isOpen) return;
-
-    function handleKeyDown(e: KeyboardEvent) {
-      if (e.key === 'Escape') onClose();
-    }
-
-    addEventListener('keydown', handleKeyDown);
-
-    return () => {
-      removeEventListener('keydown', handleKeyDown);
-    };
-  }, [isOpen]);
+  useEscapeKeyClose(isOpen, onClose);
 
   return (
     <>
