+++ components/src/lib/hooks/useEscapePress.ts
@@ -1,0 +1,19 @@
+import { useEffect } from 'react';
+
+const useEscapeKeyClose = (isOpen: boolean, onClose: () => void) => {
+  useEffect(() => {
+    if (!isOpen) return;
+
+    function handleKeyDown(e: KeyboardEvent) {
+      if (e.key === 'Escape') onClose();
+    }
+
+    addEventListener('keydown', handleKeyDown);
+
+    return () => {
+      removeEventListener('keydown', handleKeyDown);
+    };
+  }, [isOpen]);
+};
+
+export default useEscapeKeyClose;
